Player page optimization and refactoring
